# services/api/openapi.proposal.yaml
...
openapi: 3.1.0
info:
  title: TuneDeck API
  version: 0.2.0-proposal
servers:
  - url: /
paths:
...
  /v1/directory/radio:
    get:
      summary: Search the community radio directory (no sign-in)
      description: 'Radio Browser (www.radio-browser.info) searched through our server, decided by Tar 2026-10-05. These stations are community data, not our curated catalog: the apps show them as such and only add one to the user''s own list when the user picks it. Only plain public https streams (the same rules as streamUrl in StationDraft) in mp3, aac or hls get through, ordered by Radio Browser click count; stations staff blocked never appear. Answers are cached 10 minutes, and served up to an hour old while Radio Browser is down. Search text is never logged. Off (503) unless RADIO_BROWSER_BASE_URL is set; the app switch is features.radioDirectory in GET /v1/config.'
      security: []
      parameters:
        - { name: q, in: query, schema: { type: string, minLength: 1, maxLength: 80 }, description: Part of the station name; trimmed, blank is 400 }
        - { name: country, in: query, schema: { type: string, pattern: '^[A-Za-z]{2}$' }, description: ISO 3166 alpha-2, any case }
        - { name: language, in: query, schema: { type: string, minLength: 2, maxLength: 40 }, description: 'Language name as Radio Browser has it, e.g. thai' }
        - { name: tag, in: query, schema: { type: string, minLength: 1, maxLength: 30 }, description: 'Genre tag, e.g. jazz' }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 30 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, maximum: 1000, default: 0 }, description: From nextOffset }
      responses:
        '200':
          description: 'Stations (a page may hold fewer than limit after filtering; keep paging while nextOffset is set)'
          headers: { Cache-Control: { schema: { type: string, example: 'public, max-age=300' } } }
          content:
            application/json:
              schema:
                type: object
                required: [stations, nextOffset, attribution]
                properties:
                  stations: { type: array, maxItems: 50, items: { $ref: '#/components/schemas/DirectoryStation' } }
                  nextOffset: { type: integer, nullable: true }
                  attribution: { type: string, description: Show this wherever the results are shown }
        '400': { description: 'VALIDATION_FAILED: details.field q, country, language, tag, limit or offset; unknown_field for any other parameter', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '429':
          description: API_RATE_LIMITED (60 a minute per client address), details.retryAfterSeconds
          headers: { Retry-After: { schema: { type: integer } } }
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '503': { description: 'DEPENDENCY_UNAVAILABLE with details {dependency: radio_directory, reason: not_configured | unreachable}', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/admin/directory/blocks:
    get:
      summary: Community stations taken out of the directory search (catalog_editor, admin)
      description: 'Roles: catalog_editor or admin. Newest first. Cache-Control: no-store.'
      responses:
        '200':
          description: Blocks
          headers: { Cache-Control: { schema: { type: string, example: no-store } } }
          content: { application/json: { schema: { type: object, required: [blocks], properties: { blocks: { type: array, items: { $ref: '#/components/schemas/DirectoryBlock' } } } } } }
        '401': { description: 'AUTH_REQUIRED, or MFA_REQUIRED with details {scope: session, maxAgeSeconds: 43200}', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: 'ROLE_REQUIRED: needs catalog_editor or admin', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    post:
      summary: Block a community station or a stream host (catalog_editor, admin)
      description: 'Roles: catalog_editor or admin. kind station blocks one Radio Browser station id; kind host blocks every stream on that host and its subdomains. Takes effect on the next search. Audited as directory.block.add with the reason. Stations users already added stay in their own lists (Doc 10: we do not delete user libraries).'
      parameters:
        - { name: Idempotency-Key, in: header, required: false, description: 'Optional; see POST /v1/admin/stations', schema: { type: string, pattern: '^[A-Za-z0-9_.:-]{8,128}$' } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [kind, value, reason]
              properties:
                kind: { type: string, enum: [station, host] }
                value: { type: string, description: 'A Radio Browser station UUID, or a host name such as stream.example.com (lower-cased; a trailing dot is dropped)' }
                reason: { type: string, minLength: 10, maxLength: 500 }
      responses:
        '201':
          description: Blocked
          headers: { Cache-Control: { schema: { type: string, example: no-store } } }
          content: { application/json: { schema: { $ref: '#/components/schemas/DirectoryBlock' } } }
        '400': { description: 'VALIDATION_FAILED: kind value_not_allowed, value must_be_uuid or must_be_host, reason length, unknown_field', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: 'AUTH_REQUIRED, or MFA_REQUIRED with details {scope: session, maxAgeSeconds: 43200}', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: 'ROLE_REQUIRED: needs catalog_editor or admin', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: 'DIRECTORY_BLOCK_EXISTS, or IDEMPOTENCY_KEY_REUSED / IDEMPOTENCY_IN_PROGRESS', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '413': { $ref: '#/components/responses/Error' }
  /v1/admin/directory/blocks/{blockId}/remove:
    parameters:
      - { name: blockId, in: path, required: true, schema: { type: string, pattern: '^[1-9][0-9]{0,17}$' }, description: 'Not a number: 400 (field id, must_be_number)' }
    post:
      summary: Lift a block with a reason (catalog_editor, admin)
      description: 'Roles: catalog_editor or admin. Audited as directory.block.remove with the reason.'
      parameters:
        - { name: Idempotency-Key, in: header, required: false, description: 'Optional; see POST /v1/admin/stations', schema: { type: string, pattern: '^[A-Za-z0-9_.:-]{8,128}$' } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AdminExportReason' } } } }
      responses:
        '204': { description: Lifted }
        '400': { $ref: '#/components/responses/Error' }
        '401': { description: 'AUTH_REQUIRED, or MFA_REQUIRED with details {scope: session, maxAgeSeconds: 43200}', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: 'ROLE_REQUIRED: needs catalog_editor or admin', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { $ref: '#/components/responses/Error' }
        '409': { description: 'IDEMPOTENCY_KEY_REUSED or IDEMPOTENCY_IN_PROGRESS', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '413': { $ref: '#/components/responses/Error' }
...
security:
  - bearer: []
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT, description: 'OIDC access token (iss/aud/exp/RS256 verified via JWKS)' }
...
  schemas:
...
    Error:
      type: object
      required: [code, messageKey, requestId, details]
      properties:
        code:
          type: string
          enum: [AUTH_REQUIRED, AUTH_FORBIDDEN, VALIDATION_FAILED, NOT_FOUND, EXPORT_NOT_READY, IDEMPOTENCY_KEY_REQUIRED, PRECONDITION_REQUIRED, REVISION_MISMATCH, PAYLOAD_TOO_LARGE, DEPENDENCY_UNAVAILABLE, REAUTH_REQUIRED, DEVICE_REVOKED, DEVICE_LIMIT, ROLE_REQUIRED, PUBLISH_BLOCKED, CHECK_TOO_SOON, CHECKER_UNAVAILABLE, RIGHTS_ALREADY_REVOKED, API_RATE_LIMITED, ACCOUNT_DELETING, JOB_RETRY_TOO_SOON, JOB_BUSY, JOB_RETRY_LIMIT, JOB_SUPERSEDED, SYNC_RESET_REQUIRED, PURCHASE_INVALID, PURCHASE_CONFLICT, MFA_REQUIRED, IDEMPOTENCY_KEY_REUSED, IDEMPOTENCY_IN_PROGRESS, SUPPORT_ACCESS_REQUIRED, INTERNAL, DIRECTORY_BLOCK_EXISTS]
        messageKey: { type: string, example: errors.settings.revisionMismatch }
        requestId: { type: string }
        details:
          type: object
          description: 'No tokens or PII. VALIDATION_FAILED → {field?, reason, allowed?}; REVISION_MISMATCH → {currentRevision}; REAUTH_REQUIRED → {maxAgeSeconds}; MFA_REQUIRED → {maxAgeSeconds, scope?}; DEVICE_LIMIT → {maxActiveDevices}; PUBLISH_BLOCKED → {reasons}; IDEMPOTENCY_IN_PROGRESS / CHECK_TOO_SOON / JOB_RETRY_TOO_SOON → {retryAfterSeconds}.'
...
    DirectoryStation:
      type: object
      additionalProperties: false
      required: [id, name, country, language, genres, streamUrl, codec, bitrateKbps, logoUrl, homepageUrl]
      description: A community station from Radio Browser. Not part of the curated catalog and never published by us.
      properties:
        id: { type: string, format: uuid, description: Radio Browser stationuuid }
        name: { type: string, maxLength: 80 }
        country: { type: string, pattern: '^[A-Z]{2}$', nullable: true }
        language: { type: string, maxLength: 40, nullable: true }
        genres: { type: array, maxItems: 5, items: { type: string, maxLength: 30 } }
        streamUrl: { type: string, format: uri, description: Plain public https URL }
        codec: { type: string, enum: [mp3, aac, hls] }
        bitrateKbps: { type: integer, minimum: 1, maximum: 1000, nullable: true }
        logoUrl: { type: string, format: uri, nullable: true, description: 'https only; the station''s own logo, rights not checked by us' }
        homepageUrl: { type: string, format: uri, nullable: true }
    DirectoryBlock:
      type: object
      additionalProperties: false
      required: [id, kind, value, reason, createdBy, createdAt]
      properties:
        id: { type: string, pattern: '^[1-9][0-9]*$' }
        kind: { type: string, enum: [station, host] }
        value: { type: string }
        reason: { type: string }
        createdBy: { type: string, nullable: true, description: OIDC subject of the staff member }
        createdAt: { type: string, format: date-time }
...
    AdminExportReason:
      type: object
      required: [reason]
      description: 'Reason for an export, a job retry and similar recorded actions. Trimmed, then 10 to 500 characters with no control characters. Other fields are ignored.'
      properties:
        reason: { type: string, minLength: 10, maxLength: 500 }
...
  responses:
    Error:
      description: Error envelope
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }